fix(firmware): close the pre-merge review's findings [SOL-PLANT-STRUCTURE-SEAM-FIRST-STRUCTURE.C4, .C9]
A cold review of the whole branch caught a real defect in the single-precision
conversion, which had never been reviewed on its own.

Refusing a parameter on the range flag alone was wrong. That flag is raised for
any value below the smallest normal one, not only for a value that arrives as
zero, so a coefficient like 1e-38 -- fully usable, and inside the range its own
structure declares admissible -- was refused, with a diagnostic reporting a
range the value satisfies. The test is now on what arrived rather than on the
flag, and the refusal reports the zero or the infinity that makes the value
unusable rather than a number it never held.

That path was covered by nothing: deleting it outright left the suite green.
Both directions are now tested -- a value collapsing to zero is refused, a value
merely costing precision is accepted -- and each fails if the other's behaviour
is adopted.

The conversion also left a wider quantity behind in a test, passed to the seam
where it takes a narrower one. It was harmless only because the two calls
around it assert a refusal and return before writing; inverting either would
have written four bytes into an eight-byte object. The test build applies
neither the strict warnings nor the analysis, so nothing was going to say so.

Nothing asserted the single-precision property at all -- reverting the seam to
the wider type passed every check. The strict settings now refuse a silent
promotion, which is what an unsuffixed literal produces, and the analysis check
requires that setting alongside the others. This is the property that keeps the
choice of controller open, so the build is where it belongs.

The tool suite's entry point sat mid-file, so running it directly skipped every
class below it -- including all the coverage for direct calls and for the host
tier's analysis -- and still reported success. The vocabulary-only header mode
had no test; it now has four, one per condition it must still refuse.

+++ firmware/src/app/native/main.c
@@ -194,7 +194,9 @@ static void exercise_plant(const char *parameter_path)
         float value = 0.0f;
         expect(plant_model_quantity(&model, (plant_quantity_t)quantity, &value),
                "a quantity the model exposes could not be read");
-        (void)printf("plant quantity %d = %.6f\n", quantity, value);
+        /* The cast is explicit because a variadic argument is promoted
+         * whatever the source says, and the build refuses a silent one. */
+        (void)printf("plant quantity %d = %.6f\n", quantity, (double)value);
     }
 
     float unused = 0.0f;
